refactor: g5endpoint, remove auth-scheme and update the request-dto t…

+++ src/Core/Application/Generic/PenomyAPI.App.FeatG49/OtherHandlers/G49StarRateResponse.cs
@@ -5,5 +5,5 @@ namespace PenomyAPI.App.FeatG49.OtherHandlers;
 public class G49StarRateResponse : IFeatureResponse
 {
     public G49ResponseStatusCode AppCode { get; set; }
-    public long StarRate { get; set; }
+    public byte StarRate { get; set; }
 }
