+++ src/External/Presentation/PenomyAPI.Presentation.FastEndpointBasedApi/Endpoints/G49/DTOs/G49GetStarRateResponseDto.cs
@@ -2,5 +2,5 @@ namespace PenomyAPI.Presentation.FastEndpointBasedApi.Endpoints.G49.DTOs;
 
 public class G49GetStarRateResponseDto
 {
-    public long CurrentUserRate { get; set; }
+    public byte CurrentUserRate { get; set; }
 }
